Final commit
Added the div with the id of cart on the html file. On the order javascript file Added the the array has an index of the id numbers of the items purchased, the array for the item description, item price, item Qty. Making all the strings show up in the html file made the javascript show up on the page. Added the total to where the value of the orderTotal value and the dollar sign will show up in front of the total. Added the value of orderTotal to the itemCost and to keep a running total of the total cost of the customer order. Added the value of the item cost is item price to multiply the item quantity. Made the item description show up for each item, the item will be the current items price from the price array, and the item quantity will be the current value from the quantity array. the for loop has the initial value of zero, there are 4 items and its adding by one.  Added the table for the item content. The initial value for the total of the order to zero. Ready for review